+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.15.6 (tbd)
 
-- (tbd)
+- Fixed issue in `piral-cli` using the bundler without standard input (#575)
 
 ## 0.15.5 (January 23, 2023)